refactor(types): Adjust Session types
