+++ src/app/[locale]/components/add/components/ComponentAddSummary.tsx
@@ -15,24 +15,22 @@ import { useState } from 'react'
 import { useTranslations } from 'next-intl'
 import { useRouter } from 'next/navigation'
 
-import AddKeyValueComponent from '@/components/AddKeyValue'
-import AddAdditionalRolesComponent from '@/components/AddAdditionalRoles'
-import { SearchUsersModal } from '@/components/sw360'
-import CommonTabIds from '@/object-types/enums/CommonTabsIds'
-import ComponentPayload from '@/object-types/ComponentPayLoad'
-import { SideBar } from '@/components/sw360'
-import DocumentTypes from '@/object-types/enums/DocumentTypes'
+import { ApiUtils, CommonUtils } from '@/utils'
 import { COMMON_NAMESPACE } from '@/object-types/Constants'
-import Vendor from '@/object-types/Vendor'
+import { Component, InputKeyValue, HttpStatus, Session } from '@/object-types'
+import { SearchUsersModal, SideBar } from 'next-sw360'
+import AddAdditionalRolesComponent from '@/components/AddAdditionalRoles'
+import AddKeyValueComponent from '@/components/AddKeyValue'
+import CommonTabIds from '@/object-types/enums/CommonTabsIds'
 import ComponentOwner from '@/object-types/ComponentOwner'
-import Moderators from '@/object-types/Moderators'
+import ComponentPayload from '@/object-types/ComponentPayLoad'
+import DocumentTypes from '@/object-types/enums/DocumentTypes'
 import GeneralInfoComponent from '@/components/GeneralInfoComponent/GeneralInfoComponent'
+import Moderators from '@/object-types/Moderators'
 import RolesInformation from '@/components/RolesInformationComponent/RolesInformation'
 import ToastData from '@/object-types/ToastData'
 import ToastMessage from '@/components/sw360/ToastContainer/Toast'
-
-import { ApiUtils, CommonUtils } from '@/utils'
-import { Component, InputKeyValue, HttpStatus, Session } from '@/object-types'
+import Vendor from '@/object-types/Vendor'
 interface Props {
     session: Session
 }
