+++ src/app/[locale]/projects/add/components/ProjectAddSummary.tsx
@@ -21,7 +21,7 @@ import { ToastContainer } from 'react-bootstrap'
 import ApiUtils from '@/utils/api/api.util'
 import ProjectPayload from '@/object-types/CreateProjectPayload'
 import { useState } from 'react'
-import { Session } from '@/object-types/Session'
+import { Session } from '@/object-types'
 import Vendor from '@/object-types/Vendor'
 import ToastMessage from '@/components/sw360/ToastContainer/Toast'
 import ToastData from '@/object-types/ToastData'
