+++ src/app/[locale]/search/page.tsx
@@ -9,7 +9,7 @@
 
 import { authOptions } from '@/app/api/auth/[...nextauth]/route'
 import { getServerSession } from 'next-auth/next'
-import { Session } from '@/object-types/Session'
+import { Session } from '@/object-types'
 import SearchPage from './components/SearchPage'
 import { Metadata } from 'next'
